Handle OAM DMA transfer
This is implemented as a braindead loop of copying one byte at a time.
This should be fixed later. Also, it's implemented as part of the CPU
instead of the PPU and I don't know how I feel about that.

Also fixup interrupts (hardware clears the interrupt flag when triggering
the ISR).